requirements.txt
dnspython
requests
PyYAML
schedule
sqlite3
flask
jinja2
